Model/data/requirements.txt
pandas==2.2.2
google-generativeai==0.5.4
python-dotenv==1.0.1